modified moduleDependencies-test.js to conform to style guide
